requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.1.0
celery==5.2.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.7.0
Django==4.2
django-redis==5.4.0
django-role-permissions==3.2.0
django-user-agents==0.4.0
flower==1.1.0
fonttools==4.47.2
frozenlist==1.4.1
geoip2==4.8.0
html5lib==1.1
humanize==4.9.0
idna==3.6
kombu==5.3.5
maxminddb==2.5.2
multidict==6.0.4
Pillow==9.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
ua-parser==0.18.0
urllib3==2.1.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
xlwings==0.30.13
yarl==1.9.4
zopfli==0.2.3
